requirements.txt
aiofiles==0.5.0
autopep8==1.5.2
beautifulsoup4==4.9.1
bs4==0.0.1
importlib-metadata==1.6.0
Markdown==3.2.2
Pillow==7.2.0
pycodestyle==2.6.0
python-dateutil==2.8.1
PyYAML==5.3.1
redis==3.5.2
rope==0.17.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
zipp==3.1.0
